refector: 패키지 의존관계 확인 및 rename
