+++ README.md
@@ -43,8 +43,8 @@
 
 ## 5단계 - 자동차 경주(리팩토링)
 
-- [ ] 핵심 비지니스 로직을 가지는 객체를 domain 패키지, UI 관련한 객체를 view 패키지에 구현
-- [ ] view 패키지의 객체가 domain 패키지 객체에 의존할 수 있지만, domain 패키지의 객체는 view 패키지 객체에 의존하지 않도록 구현
+- [x] 핵심 비지니스 로직을 가지는 객체를 domain 패키지, UI 관련한 객체를 view 패키지에 구현
+- [x] view 패키지의 객체가 domain 패키지 객체에 의존할 수 있지만, domain 패키지의 객체는 view 패키지 객체에 의존하지 않도록 구현
 - [ ] 테스트 가능한 부분에 대해서만 단위 테스트를 진행
 - [ ] 규칙 3: 모든 원시값과 문자열을 포장한다.
 - [ ] 규칙 8: 일급 콜렉션을 쓴다.
