+++ src/main/java/carrace/view/input/scanner/InputScanner.java
@@ -1,4 +1,4 @@
-package carrace.ui.input.scanner;
+package carrace.view.input.scanner;
 
 public interface InputScanner<T> {
     T scan();
