+++ src/main/java/carrace/view/input/scanner/NameInputScanner.java
@@ -1,4 +1,4 @@
-package carrace.ui.input.scanner;
+package carrace.view.input.scanner;
 
 import java.util.NoSuchElementException;
 import java.util.Scanner;
